--- driver/js/src/js_driver_utils_any_cast.cc
@@ -1,38 +1,0 @@
-/*
- * Tencent is pleased to support the open source community by making
- * Hippy available.
- *
- * Copyright (C) 2022 THL A29 Limited, a Tencent company.
- * All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#include "driver/js_driver_utils.h"
-#include <any>
-
-namespace hippy {
-inline namespace driver {
-
-// 公共的any_cast工具方法
-ScopeWrapper* JsDriverUtils::GetScopeWrapperFromSlot(const std::any& slot) {
-  auto slot_ptr = std::any_cast<void*>(&slot);
-  if (slot_ptr) {
-    return static_cast<ScopeWrapper*>(*slot_ptr);
-  }
-
-  return nullptr;
-}
-
-}  // namespace driver
-}  // namespace hippy
